requirements.txt
Flask==2.1.2
gunicorn==20.1.0
python-dotenv==0.20.0
requests==2.28.1
pymongo[tls,srv]==4.1.1
flask-marshmallow==0.14.0